Move view file to subdirectory
