+++ lib/sync-sync.js
@@ -1,6 +1,6 @@
 'use babel';
 
-import SyncSyncView from './sync-sync-view';
+import SyncSyncView from './views/sync-sync-view';
 import { CompositeDisposable } from 'atom';
 
 export default {
