+++ lib/views/sync-sync-view.js
